Crate sgx_tprotected_fs

Macros

eos

esgx

impl_asmut_slice

impl_asref_slice

Structs

EncryptMode

OpenOptions

Options and flags which can be used to configure how a file is opened.

SgxFile

A reference to an open Sgxfile on the filesystem.

Functions

read_integrity_only

read_to_string_integrity_only

read_to_string_with_key

read_with_key

remove

write_integrity_only

write_with_key