Function sgx_tprotected_fs::import_auto_key

pub fn import_auto_key(filename: &CStr, key: &sgx_key_128bit_t) -> SysError

The import_auto_key function is used for importing a Protected FS auto key file created on a different enclave or platform.

Description

import_auto_key is used for importing a Protected FS file. After this call returns successfully, the file can be opened normally with export_auto_key.

Parameters

filename

The name of the file to be imported. This should be the name of a file created with the open_auto_key API, on a different enclave or system.

key

he encryption key, exported with a call to export_auto_key in the source enclave or system.

Requirements

Header: sgx_tprotected_fs.edl

Library: libsgx_tprotected_fs.a

Return value

otherwise, error code is returned.