Struct sgx_tprotected_fs::SgxFile

pub struct SgxFile { /* private fields */ }

A reference to an open Sgxfile on the filesystem.

An instance of a SgxFile can be read and/or written depending on what options it was opened with. SgxFiles also implement [Seek] to alter the logical cursor that the file contains internally.

SgxFiles are automatically closed when they go out of scope.

Implementations

impl SgxFile

pub fn open<P: AsRef<Path>>(path: P) -> Result<SgxFile>

pub fn create<P: AsRef<Path>>(path: P) -> Result<SgxFile>

pub fn append<P: AsRef<Path>>(path: P) -> Result<SgxFile>

pub fn open_with_key<P: AsRef<Path>>(path: P, key: Key128bit) -> Result<SgxFile>

pub fn create_with_key<P: AsRef<Path>>(
path: P,
key: Key128bit
) -> Result<SgxFile>

pub fn append_with_key<P: AsRef<Path>>(
path: P,
key: Key128bit
) -> Result<SgxFile>

pub fn open_integrity_only<P: AsRef<Path>>(path: P) -> Result<SgxFile>

pub fn create_integrity_only<P: AsRef<Path>>(path: P) -> Result<SgxFile>

pub fn append_integrity_only<P: AsRef<Path>>(path: P) -> Result<SgxFile>

pub fn open_with<P: AsRef<Path>>(
path: P,
encrypt_mode: EncryptMode,
cache_size: Option<usize>
) -> Result<SgxFile>

pub fn create_with<P: AsRef<Path>>(
path: P,
encrypt_mode: EncryptMode,
cache_size: Option<usize>
) -> Result<SgxFile>

pub fn append_with<P: AsRef<Path>>(
path: P,
encrypt_mode: EncryptMode,
cache_size: Option<usize>
) -> Result<SgxFile>

pub fn options() -> OpenOptions

pub fn set_len(&self, size: u64) -> Result<()>

pub fn tell(&self) -> Result<u64>

pub fn file_size(&self) -> Result<u64>

pub fn is_eof(&self) -> bool

pub fn clear_error(&self) -> Result<()>

pub fn clear_cache(&self) -> Result<()>

pub fn get_mac(&self) -> Result<Mac128bit>

pub fn rename<P: AsRef<str>>(&mut self, old_name: P, new_name: P) -> Result<()>