Crate sgx_tprotected_fs

Intel Protected File System Library

Intel Protected File System Library provides protected files API for Intel SGX enclaves. With this API, the files are encrypted and saved on the untrusted disk during a write operation, and they are verified for confidentiality and integrity during a read operation.

Structs

SgxFileStream

Enums

SeekFrom
Enumeration of possible methods to seek within an I/O object.

Functions

export_align_auto_key

export_auto_key
The export_auto_key function is used for exporting the latest key used for the file encryption.

import_auto_key
The import_auto_key function is used for importing a Protected FS auto key file created on a different enclave or platform.

remove
The remove function deletes a file from the file system.